server_python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.1
openai==1.3.5
apscheduler==3.10.4
logfire[fastapi]==0.51.0
